Introduce separate slug field, add author to articles improve article categories.

+++ src/utils/format-datetime.ts
@@ -1,5 +1,5 @@
 const formatter = Intl.DateTimeFormat("de-DE", {
-  dateStyle: "long",
+  dateStyle: "short",
   timeStyle: "short",
 });
